requirements.txt
python-telegram-bot==20.8
httpx==0.27.2
certifi==2024.12.14
httpcore==1.0.7
idna==3.10
sniffio==1.3.1
python-dotenv==1.0.0